referendum code update
